requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.35
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
Pillow>=10.0
tqdm>=4.65
matplotlib>=3.7
grad-cam>=1.5
git+https://github.com/openai/CLIP.git
